remove canvas wrapper
