+++ src/lib/Shape.ts
@@ -1,0 +1,5 @@
+export interface Shape {
+    draw(ctx: CanvasRenderingContext2D): void;
+
+    handleClick(x: number, y: number): void;
+}
